Move Tasks into Duke package
