+++ src/main/java/duke/BaseTask.java
@@ -1,6 +1,4 @@
-package duke.Tasks;
-
-import duke.DukeExceptionBase;
+package duke;
 
 
 /**
