+++ src/main/java/duke/DeadlineTask.java
@@ -1,6 +1,4 @@
-package duke.Tasks;
-
-import duke.DukeExceptionBase;
+package duke;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
